dedupe kafka summaries
